+++ app/service/entrypoint.py
@@ -49,6 +49,7 @@ def main():
     keep.update(HOME="/home/browser", DOUBLETAKE_BROWSER="/opt/browser-app/acceleration.py",
                 DOUBLETAKE_BROWSER_CONTROL=options.get('browser_control', 'native'),
                 DOUBLETAKE_AUDIO="true" if options.get("audio", True) else "false",
+                DOUBLETAKE_HARDWARE_ENCODING="true" if options.get("hardware_encoding", True) else "false",
                 DOUBLETAKE_HARDWARE_DECODING="true" if options.get("hardware_decoding", True) else "false")
     os.environ.clear()
     os.environ.update(keep)